backend/services/video_merger/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
ffmpeg-python==0.2.1
webvtt-py==0.5.1
python-dotenv==1.0.0
redis==5.0.1
psutil==5.9.6
aiofiles==23.2.1
